# Improve performance of Fiberrefs#join