resolves #6853 
```
Before:

FiberRefBenchmarks.createAndJoin               32      10000  thrpt   10  27,539 ± 0,718  ops/s
FiberRefBenchmarks.createAndJoinExpensive   32     10000  thrpt   10  2,718 ± 0,039  ops/s
FiberRefBenchmarks.createAndJoinInitialValue   32      10000  thrpt   10  27,621 ± 1,142  ops/s
FiberRefBenchmarks.createAndJoinUpdatesDeep    32      10000  thrpt   10   8,732 ± 0,532  ops/s
FiberRefBenchmarks.createAndJoinUpdatesWide    32      10000  thrpt   10  10,727 ± 0,127  ops/s

After:

FiberRefBenchmarks.createAndJoin               32      10000  thrpt   10  53,797 ± 1,778  ops/s
FiberRefBenchmarks.createAndJoinExpensive   32 10000  thrpt   10  41,947 ± 1,707  ops/s
FiberRefBenchmarks.createAndJoinInitialValue   32      10000  thrpt   10  53,137 ± 1,670  ops/s
FiberRefBenchmarks.createAndJoinUpdatesDeep    32      10000  thrpt   10   8,296 ± 0,364  ops/s
FiberRefBenchmarks.createAndJoinUpdatesWide    32      10000  thrpt   10  11,375 ± 0,175  ops/s
```

No microoptimizations, just short cicuiting when we can avoid calculating the patches.
As seen in `createAndJoinExpensive` this is has the strongest effect with Fiberrefs that are expensive to diff or upgrade.